Split src/viewer/index.js to distinct parts
This is a partial step in refactoring the whole, as it allows me to
focus on less things at once.

I don't know yet the complete file structure of what the source will
look like in the end, but this is a useful, safe refactoring step on the
way to get there :smile:
